requirements.txt
alembic==1.12.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
dominate==2.8.0
email-validator==2.1.0.post1
Faker==20.1.0
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mimesis==11.1.0
opensearch-py==2.3.2
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.0
